backend/requirements.txt
Django>=4.2
djangorestframework
django-cors-headers
djangorestframework-simplejwt
argon2-cffi
google-auth
google-auth-httplib2
requests
python-dotenv
channels[daphne]
channels
PyJWT[crypto]
cryptography
